Revert "Revert "[JAVA-8154] Code clean up for reactive security""
