--- spring-reactive/src/main/java/com/baeldung/reactive/security/GreetController.java
@@ -1,37 +1,0 @@
-package com.baeldung.reactive.security;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import reactor.core.publisher.Mono;
-
-import java.security.Principal;
-
-@RestController
-public class GreetController {
-
-    private GreetService greetService;
-
-    public GreetController(GreetService greetService) {
-        this.greetService = greetService;
-    }
-
-    @GetMapping("/")
-    public Mono<String> greet(Mono<Principal> principal) {
-        return principal
-                .map(Principal::getName)
-                .map(name -> String.format("Hello, %s", name));
-    }
-
-    @GetMapping("/admin")
-    public Mono<String> greetAdmin(Mono<Principal> principal) {
-        return principal
-                .map(Principal::getName)
-                .map(name -> String.format("Admin access: %s", name));
-    }
-
-    @GetMapping("/greetService")
-    public Mono<String> greetService() {
-        return greetService.greet();
-    }
-
-}
