+++ spring-reactive/src/main/java/com/baeldung/reactive/security/SecurityConfig.java
@@ -16,40 +16,37 @@ import org.springframework.security.web.server.SecurityWebFilterChain;
 public class SecurityConfig {
 
     @Bean
-    public SecurityWebFilterChain securitygWebFilterChain(ServerHttpSecurity http) {
+    public SecurityWebFilterChain securityWebFilterChain(ServerHttpSecurity http) {
         return http.authorizeExchange()
-            .pathMatchers("/admin")
-            .hasAuthority("ROLE_ADMIN")
-            .anyExchange()
-            .authenticated()
-            .and()
-            .formLogin()
-            .and()
-            .csrf()
-            .disable()
-            .build();
+          .pathMatchers("/admin").hasAuthority("ROLE_ADMIN")
+          .anyExchange().authenticated()
+          .and()
+          .formLogin()
+          .and()
+          .csrf().disable()
+          .build();
     }
 
     @Bean
     public MapReactiveUserDetailsService userDetailsService() {
         UserDetails user = User
-            .withUsername("user")
-            .password(passwordEncoder().encode("password"))
-            .roles("USER")
-            .build();
+          .withUsername("user")
+          .password(passwordEncoder().encode("password"))
+          .roles("USER")
+          .build();
 
         UserDetails admin = User
-            .withUsername("admin")
-            .password(passwordEncoder().encode("password"))
-            .roles("ADMIN")
-            .build();
+          .withUsername("admin")
+          .password(passwordEncoder().encode("password"))
+          .roles("ADMIN")
+          .build();
 
         return new MapReactiveUserDetailsService(user, admin);
     }
-    
+
     @Bean
     public PasswordEncoder passwordEncoder() {
-    	return new BCryptPasswordEncoder();
+        return new BCryptPasswordEncoder();
     }
 
 }
